+++ src/main/java/com/example/springboard/dto/request/UserRequest.java
@@ -1,4 +1,4 @@
-package com.example.springboard.dto;
+package com.example.springboard.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
