+++ src/main/java/com/example/springboard/dto/request/UserSignupRequest.java
@@ -1,29 +1,26 @@
-package com.example.springboard.dto;
+package com.example.springboard.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Pattern;
 
-public class UserUpdateRequest {
+public class UserSignupRequest {
+
+    @NotBlank(message = "아이디를 입력해주세요")
     private String userId;
 
-    @NotBlank(message = "변경할 비밀번호를 입력해주세요")
+    @NotBlank(message = "비밀번호를 입력해주세요")
     @Pattern(regexp = "^(?=.*[A-Za-z])(?=.*\\d)(?=.*[@$!%*#?&])[A-Za-z\\d@$!%*#?&]{8,30}$",
             message = "비밀번호는 8~30 자리이면서 1개 이상의 알파벳, 숫자, 특수문자를 포함해야합니다.")
-    private String newPassword;
-
     private String password;
 
+    @NotBlank(message = "닉네임을 입력해주세요")
     private String nickname;
 
-    public UserUpdateRequest() {
+    public UserSignupRequest() {
     }
 
     public String getUserId() {
         return userId;
-    }
-
-    public String getNewPassword() {
-        return newPassword;
     }
 
     public String getPassword() {
@@ -36,9 +33,8 @@ public class UserUpdateRequest {
 
     @Override
     public String toString() {
-        return "UserUpdateRequest{" +
+        return "UserSignupRequest{" +
                 "userId='" + userId + '\'' +
-                ", newPassword='" + newPassword + '\'' +
                 ", password='" + password + '\'' +
                 ", nickname='" + nickname + '\'' +
                 '}';
