+++ src/main/java/com/example/springboard/dto/request/UserUpdateRequest.java
@@ -1,26 +1,29 @@
-package com.example.springboard.dto;
+package com.example.springboard.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Pattern;
 
-public class UserSignupRequest {
-
-    @NotBlank(message = "아이디를 입력해주세요")
+public class UserUpdateRequest {
     private String userId;
 
-    @NotBlank(message = "비밀번호를 입력해주세요")
+    @NotBlank(message = "변경할 비밀번호를 입력해주세요")
     @Pattern(regexp = "^(?=.*[A-Za-z])(?=.*\\d)(?=.*[@$!%*#?&])[A-Za-z\\d@$!%*#?&]{8,30}$",
             message = "비밀번호는 8~30 자리이면서 1개 이상의 알파벳, 숫자, 특수문자를 포함해야합니다.")
+    private String newPassword;
+
     private String password;
 
-    @NotBlank(message = "닉네임을 입력해주세요")
     private String nickname;
 
-    public UserSignupRequest() {
+    public UserUpdateRequest() {
     }
 
     public String getUserId() {
         return userId;
+    }
+
+    public String getNewPassword() {
+        return newPassword;
     }
 
     public String getPassword() {
@@ -33,8 +36,9 @@ public class UserSignupRequest {
 
     @Override
     public String toString() {
-        return "UserSignupRequest{" +
+        return "UserUpdateRequest{" +
                 "userId='" + userId + '\'' +
+                ", newPassword='" + newPassword + '\'' +
                 ", password='" + password + '\'' +
                 ", nickname='" + nickname + '\'' +
                 '}';
