+++ src/main/java/com/example/springboard/mapper/UserMapper.java
@@ -1,9 +1,9 @@
 package com.example.springboard.mapper;
 
-import com.example.springboard.dto.UserSignupRequest;
-import com.example.springboard.dto.UserUpdateRequest;
+import com.example.springboard.dto.request.UserSignupRequest;
+import com.example.springboard.dto.request.UserUpdateRequest;
 import com.example.springboard.vo.User;
-import com.example.springboard.dto.UserRequest;
+import com.example.springboard.dto.request.UserRequest;
 import org.apache.ibatis.annotations.Mapper;
 
 @Mapper
