+++ src/main/resources/mapper/BoardMapper.xml
@@ -4,7 +4,7 @@
 
 <mapper namespace="com.example.springboard.mapper.BoardMapper">
     <!--게시글 목록 불러오기 -->
-    <select id="getPostList" resultType="com.example.springboard.dto.PostResponse">
+    <select id="getPostList" resultType="com.example.springboard.dto.response.PostResponse">
         select a.*, b.nickname, c.comments_count
         from POST as a
         left join user as b
